refactor(TravelPlanDetailPage): 컴포넌트 명 및 디렉터리 변경
1. TravelPlansDetail에서 travelPlanDetail로 디렉터리 변경

2. TravelPlanDetailPage로 컴포넌트 네이밍 변경

3. 변경 사항 다른 컴포넌트들에 반영
