+++ frontend/src/components/pages/travelPlanDetail/TravelPlanDetailPage.styled.ts
